refactor: move icon suggestion to a general api route
