+++ test-utils/playwright/test-tables/finance.ts
@@ -1,4 +1,4 @@
-import { CategoryIconsMap } from "#shared/finance/maps/category-icons.map";
+import { IconTagsMap } from "#shared/core/maps/icon-tags.map";
 import { BASE_ACCOUNT, BASE_AUTO_TRANSACTION, BASE_TRANSACTION } from "~~/test-utils/constants/finance";
 import { str } from "~~/test-utils/helpers/data";
 import { omit } from "~~/test-utils/helpers/object";
@@ -118,7 +118,7 @@ export const TRANSACTION_VALID_REQUEST_TEST_CASES = [
         withTransaction("reference", ` ${str({ length: 100 })} `),
         withTransaction("reference", str({ length: 100 })),
     ],
-    ...(Object.keys(CategoryIconsMap).map(icon => [
+    ...(Object.keys(IconTagsMap).map(icon => [
         `the category icon is ${icon}`,
         withTransaction("category", { ...BASE_TRANSACTION.category, icon }),
     ]) as [string, typeof BASE_TRANSACTION][]),
@@ -172,7 +172,7 @@ export const AUTO_TRANSACTION_VALID_REQUEST_TEST_CASES = [
     ["execOn is just large enough", withAutoTransaction("execOn", 1)],
     ["execOn is just small enough", withAutoTransaction("execOn", 31)],
     ["lastExec is a valid date", withAutoTransaction("lastExec", "2025-06-30")],
-    ...(Object.keys(CategoryIconsMap).map(icon => [
+    ...(Object.keys(IconTagsMap).map(icon => [
         `the category icon is ${icon}`,
         withAutoTransaction("category", { ...BASE_AUTO_TRANSACTION.category, icon }),
     ]) as [string, typeof BASE_AUTO_TRANSACTION][]),
