+++ tests/unit/shared/core/maps/icon-tags.map.test.ts
@@ -1,9 +1,9 @@
+import { IconTagsMap } from "#shared/core/maps/icon-tags.map";
 import { TITLE_MAX } from "#shared/core/validators/core.validator";
-import { CategoryIconsMap } from "#shared/finance/maps/category-icons.map";
 import { expect, test } from "vitest";
 
 test(`contains no key that is longer than ${TITLE_MAX}`, () => {
-    const keys = Object.keys(CategoryIconsMap);
+    const keys = Object.keys(IconTagsMap);
     const containsLongKeys = keys.some(key => key.length > TITLE_MAX);
     expect(containsLongKeys).toBe(false);
 });
